+++ src/main/Core/BrowserWindow/BrowserWindowToggler/BrowserWindowToggler.test.ts
