+++ src/main/Core/BrowserWindow/BrowserWindowToggler/BrowserWindowToggler.ts
