+++ src/main/Core/BrowserWindow/BrowserWindowToggler/index.ts
@@ -1,0 +1,1 @@
+export * from "./BrowserWindowToggler";
